Whatsapp Chat Stats
I was curious to see how much each person in a group chat texted, how many words per text they had on average, and the number of alternative messages they had such as images, stickers, etc. It also includes a specific word counter; in this case, it is left as "nacky".

The file input is simply the .txt file generated from unzipping the file generated from clicking "export chat" on a WhatsApp chat and selecting not to include media. If done on a computer, this may be limited to a certain amount of storage, but you can get the entire chat history by exporting on your phone and sending to your computer.